+++ sms-facade/src/main/java/com/ruoyi/sms/facade/dto/SysOrderDTO.java
@@ -26,7 +26,7 @@ public class SysOrderDTO extends ToString {
     /**
      * null
      */
-    private Long id;
+    private Integer id;
 
     /**
      * 订单ID
@@ -86,7 +86,7 @@ public class SysOrderDTO extends ToString {
     /**
      * 商品信息
      */
-    private Long goodsId;
+    private Integer goodsId;
 
     /**
      * 商品快照
@@ -103,11 +103,11 @@ public class SysOrderDTO extends ToString {
      */
     private String extensionUserId;
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
@@ -183,11 +183,11 @@ public class SysOrderDTO extends ToString {
         return status;
     }
 
-    public void setGoodsId(Long goodsId) {
+    public void setGoodsId(Integer goodsId) {
         this.goodsId = goodsId;
     }
 
-    public Long getGoodsId() {
+    public Integer getGoodsId() {
         return goodsId;
     }
